[Design System] Allow Modal component to enable background scrolling (#208)
Add a disableBackgroundScroll prop to the Modal component in the design system, which is true by default but prevents background scroll from being locked when it is false.

This will enable us to implement new designs in Workflows for the opportunity side panel. This side panel uses the DrawerModal component, which inherits from Modal. We would now like the user to be able to interact with the rest of the page while the modal is open.